+++ package.json
@@ -26,6 +26,7 @@
     "jspdf-autotable": "^5.0.2",
     "lucide-react": "^0.525.0",
     "node-fetch": "^3.3.2",
+    "p-limit": "^6.2.0",
     "qs": "^6.14.0",
     "react": "^19.1.0",
     "react-chartjs-2": "^5.3.0",
